requirements.txt
numpy==1.24.2
matplotlib==3.7.0
notebook==6.5.2
scikit-learn==1.2.1
scikit-image==0.20.0
Pillow==9.4.0
Flask==2.2.3
requests==2.28.2
torch==1.13.1
torchvision==0.14.1

